Information Retrieval System/requirements.txt
  python-dotenv
  google-generativeai
  langchain
  PyPDF2
  faiss-cpu
  streamlit
  langchain-community

